item delete handler tests
